requirements.txt
google-cloud-aiplatform
Pillow
python-dotenv
google-cloud-storage